// src/app/page/master/kelompok-status/kelompok-status.interface.ts
import {EventEmitter} from '@angular/core';

export interface KelompokStatus {
	kdProfile?: string;
	kdKelompokStatus?: string;
	namaKelompokStatus?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;

 }